Separate unusable file removal from set splitting
