+++ scripts/imagenet/from_urls/remove_unusable_images.py
@@ -1,10 +1,31 @@
 #! /usr/bin/env python
-"""Split Imagnet images into train, val, and test sets"""
+"""Remove unloadable or corrupt images from the downloaded ImageNet dataset
+
+This script removes images that were corrupted during download, were
+not downloaded as image data (e.g. as HTML instead), or that are otherwise
+unloadable for some other reason.
+
+This script expects the following files to be present on disk:
+
+    1. /data/imagenet/from_urls/metadata_lists/bad_files.txt: this holds image
+       URLs that were corrupted during download and /or not downloaded as image
+       data. It can be generated by running the following at the command line:
+
+       'find /data/imagenet/from_urls/images -type f | xargs file | grep -v '
+       'tee /data/imagenet/from_urls/metadata_lists/bad_files.txt '
+
+       **Note**: This command may need to be updated if your images are stored
+       at a different location, and this may take 12+ hours to run depending on
+       how many images you have stored.
+
+    2. /data/imagenet/from_urls/metadata_lists/unloadable_files.csv: this holds
+       images that are unloadable using `imageio`, and can be generated by
+       running 'scripts/imagenet/from_urls/find_uloadable_images.py'.
+"""
 
 import os
 
 import pandas as pd
-from sklearn.model_selection import train_test_split
 
 from utils import dev_env
 
@@ -17,6 +38,9 @@ FPATH_BAD_FILES_TXT = os.path.join(DIRPATH_METADATA_LISTS, 'bad_files.txt')
 FPATH_DF_FPATHS_IMAGES = os.path.join(
     DIRPATH_METADATA_LISTS, 'df_fpaths_images.csv'
 )
+FPATH_DF_FPATHS_IMAGES_FILTERED = os.path.join(
+    DIRPATH_METADATA_LISTS, 'df_fpaths_images_filtered.csv'
+)
 FPATH_UNLOADABLE_FILES_CSV = os.path.join(
     DIRPATH_METADATA_LISTS, 'unloadable_files.csv'
 )
@@ -24,28 +48,6 @@ FPATH_UNLOADABLE_FILES_CSV = os.path.join(
 FPATH_SYNSET_WORDS = os.path.join(
     DIRPATH_DATA, 'synset_lists', 'synset_words.txt'
 )
-
-
-def add_label_column(df_fpaths_images):
-    """Add the corresponding 'label' for each 'synset' in `df_fpaths_images`
-
-    This translates the 'synset' column to an integer label for use in
-    training, i.e. n01580077 => 'jay'.
-
-    :param df_fpaths_images: holds the filepaths for the imagenet images
-    :type df_fpaths_images: pandas.DataFrame
-    :return: df_fpaths_images with 'label' column added
-    :rtype: pandas.DataFrame
-    """
-
-    df_synsets_text = pd.read_table(FPATH_SYNSET_WORDS, names=['text'])
-    synsets, _ = df_synsets_text['text'].str.split(' ', 1).str
-    df_synsets = pd.DataFrame(synsets)
-    df_synsets.rename(columns={'text': 'synset'}, inplace=True)
-    df_synsets['label'] = df_synsets.index
-
-    df_fpaths_images = pd.merge(df_fpaths_images, df_synsets, on='synset')
-    return df_fpaths_images
 
 
 def remove_bad_files(df_fpaths_images):
@@ -112,47 +114,22 @@ def main():
             'image URLs that were downloaded as BAD files instead of images, '
             'as well as the files that are unloadable. To generate {0}, type '
             'the following at the command line: '
-            'find /data/imagenet/images -type f | xargs file | grep -v '
-            r'"image data" | '
-            'tee /data/imagenet/metadata_lists/bad_files.txt | wc -l '
+            'find /data/imagenet/from_urls/images -type f | xargs file | '
+            'grep -v | "image data" | '
+            'tee /data/imagenet/from_urls/metadata_lists/bad_files.txt '
             '\n\n'
             '**Note**: This command may need to be updated if your images are '
             'stored at a different location, and this may take 12+ hours to '
             'run depending on how many images you have stored. \n\n '
             'To generate {1}, run the \'find_unloadable_imagenet_images.py\' '
-            'script in the scripts directory.'
+            'script in the scripts/imagenet/from_urls directory.'
         ).format(FPATH_BAD_FILES_TXT, FPATH_UNLOADABLE_FILES_CSV)
         raise FileNotFoundError(msg)
 
     df_fpaths_images = pd.read_csv(FPATH_DF_FPATHS_IMAGES)
     df_fpaths_images = remove_bad_files(df_fpaths_images)
     df_fpaths_images = remove_unloadable_files(df_fpaths_images)
-    df_fpaths_images = add_label_column(df_fpaths_images)
-
-    df_train, df_test = train_test_split(
-        df_fpaths_images, train_size=0.80, test_size=0.20, random_state=529,
-        stratify=df_fpaths_images['synset']
-    )
-    df_val, df_test = train_test_split(
-        df_test, train_size=0.50, test_size=0.50, random_state=529,
-        stratify=df_test['synset']
-    )
-
-    fpath_test_set = os.path.join(
-        DIRPATH_METADATA_LISTS, 'df_test_set.csv'
-    )
-    df_test.to_csv(fpath_test_set, index=False)
-
-    fpath_val_set = os.path.join(
-        DIRPATH_METADATA_LISTS, 'df_val_set.csv'
-    )
-    df_val.to_csv(fpath_val_set, index=False)
-
-    fpath_train_set = os.path.join(
-        DIRPATH_METADATA_LISTS, 'df_train_set.csv'
-    )
-    df_train.to_csv(fpath_train_set, index=False)
-
+    df_fpaths_images.to_csv(FPATH_DF_FPATHS_IMAGES_FILTERED, index=False)
 
 if __name__ == '__main__':
     main()
